requirements.txt
Pypubsub => 4.0.3
wxPython => 4.2.3